Restructure src/app/ into three-bucket layout
Move contacts-specific components into app/contacts/ and the
presence widget into app/presence/. app/components/ is reserved for
generic primitives and stays empty until first primitive migration.
Colocate the contacts-list test next to ContactsList.jsx.

Document the three-bucket rule (primitives | feature-scoped | root
composition) and PascalCase/kebab-case/Dialog-vs-modal naming in
STRUCTURE.md. Mark app-layer UI location as resolved in
SOLID_UI_STATE_RULES.md; update stale paths in SOLIDJS_POC.md.
